Moved esil from plugin and exported analysis from esil code
